+++ migrations/2021-02-25-170143_cdbs/up.sql
@@ -124,7 +124,7 @@ CREATE TABLE file_ref (
   created_at TIMESTAMP NOT NULL DEFAULT NOW(), /* дата создания/загрузки */
   filename VARCHAR(225) NOT NULL, /* наименование файла */
   id_ext INTEGER NOT NULL, /* расширение файла (используется для определения CAD) */
-  filesize FLOAT NOT NULL, /* размер файла */
+  filesize INTEGER NOT NULL, /* размер файла */
   path_file VARCHAR(225) NOT NULL, /* путь к файлу */
   CONSTRAINT file_ref_pk PRIMARY KEY (uuid)
 );
